Move iterators to detail
